+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Image Annotation Tool
 
+## 🏷 1.1.0 &#8239;&#8239;&#8239;&#8239; 🗓 2021-08-24
+
+### New Features
+* Hand tool was added to the toolbar to allow user to move the canvas on the board.
+### Changes
+* The function of move tool was changed. From now on it should be used to move marked zones on the image.
+
 ## 🏷 1.0.0 &#8239;&#8239;&#8239;&#8239; 🗓 2021-08-23
 First release of the Image Annotation Tool. The app contains basic set of tools.